# spec/openapi/openreceive-http.v1.yaml
openapi: 3.1.0
info:
  title: OpenReceive storage-free HTTP API
  version: 0.2.0
  description: >
    Mounted receive-checkout routes. OpenReceive owns no payment persistence. The host
    resolves every checkout amount and stores payment_hash before a payer response is sent.
servers:
  - url: /openreceive
paths:
...
  /rates:
    get:
      operationId: listRates
      parameters:
        - in: query
          name: currencies
          schema: { type: string }
      responses:
        '200':
          description: Current process-cached BTC/fiat rates.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
components:
  securitySchemes:
    paymentCapability:
      type: http
      scheme: bearer
      description: Opaque authenticated token bound to order_id and payment_hash.
...